Split app.js and added normalize.min.css locally
