+++ desktop-app/index.html
@@ -7,7 +7,7 @@
   <title>Tinder⁺⁺</title>
   <meta name="description" content="Desktop optimized Tinder, plus all the features of Tinder Plus for free!">
 
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/3.0.1/normalize.min.css"/>
+  <link rel="stylesheet" href="normalize.min.css"/>
   <link href='https://fonts.googleapis.com/css?family=Raleway' rel='stylesheet' type='text/css'>
   <link href='https://fonts.googleapis.com/css?family=Merriweather' rel='stylesheet' type='text/css'>
   <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.min.css" rel="stylesheet">
@@ -30,7 +30,10 @@
   <script src="js/swing.js"></script>
   <script src="js/sweet-alert.min.js"></script>
   <script src="js/twemoji.min.js"></script>
-  <script src="js/app.js"></script>
+  <script src="js/tinder++.api.js"></script>
+  <script src="js/tinder++.login.js"></script>
+  <script src="js/tinder++.swipe.js"></script>
+  <script src="js/tinder++.js"></script>
   <script src="https://maps.googleapis.com/maps/api/js?libraries=places"></script>
 </head>
 <body ng-app="tinder++">
