+++ desktop-app/js/tinder++.swipe.js
@@ -1,118 +1,8 @@
 (function() {
-  // resize to window to full screen height
-  var resizeToHeight = Math.min(820, window.screen.availHeight);
-  window.resizeTo(window.innerWidth, resizeToHeight);
+  gui = require('nw.gui')
+  module = angular.module('tinder++.swipe', ['ngAutocomplete', 'tinder++.api']);
 
-  var gui = require('nw.gui');
-  var win = gui.Window.get();
-
-  if (process.platform === 'darwin') {
-    var nativeMenuBar = new gui.Menu({ type: 'menubar' });
-    nativeMenuBar.createMacBuiltin('Tinder⁺⁺', {
-      hideEdit: true
-    });
-    win.menu = nativeMenuBar;
-  }
-
-  var app = angular.module('tinder++', ['ngAutocomplete', 'ngRoute']);
-  var tinder = require('tinderjs');
-  var client = new tinder.TinderClient();
-  if (localStorage.tinderToken) { client.setAuthToken(localStorage.tinderToken); }
-
-  app.config(function($routeProvider) {
-    ['login', 'swipe'].forEach(function(route) {
-        $routeProvider.when('/' + route, {
-            templateUrl: route + '.html',
-            controller: capitalize(route) + 'Controller'
-        });
-    });
-  });
-
-  app.run(function($location) {
-      var firstPage = (localStorage.tinderToken ? '/swipe' : '/login');
-      $location.path(firstPage);
-  });
-
-  app.factory('API', function API() {
-    var apiObj = {};
-
-    apiObj.login = function(id, token) {
-      ga_storage._trackEvent('Login', 'Facebook Login Successful');
-      client.authorize(token, id, function(err, res, data) {
-        console.log(res);
-        localStorage.tinderToken = client.getAuthToken();
-        localStorage.name = res.user.full_name;
-        localStorage.smallPhoto = res.user.photos[0].processedFiles[3].url;
-        if (window.loginWindow) {
-          window.loginWindow.close(true);
-        }
-        window.location.reload();
-      });
-    };
-    apiObj.updateLocation = function(lng, lat, callback) {
-      client.updatePosition(lng, lat, function(err, res, data) {
-        console.log(res);
-        callback();
-      });
-    };
-    apiObj.people = function(callbackFn, limit) {
-      limit = limit || 10;
-      client.getRecommendations(limit, function(err, res, data) {
-        if ((res && res.message && (res.message === 'recs timeout' || res.message === 'recs exhausted')) || !res) {
-          swal({
-            title: 'Out of people for now',
-            text: 'This can happen if you change location too much. Try quitting, opening phone app, ' +
-            'then re-opening this app to fix the problem, otherwise just wait an hour or so.',
-            type: 'error',
-            confirmButtonColor: "#DD6B55",
-            confirmButtonText: 'Got it'
-          });
-        } else {
-          if (res && res.results) {
-            callbackFn(res.results);
-          } else {
-            callbackFn([]);
-          }
-        }
-      });
-    };
-    apiObj.userInfo = function(userId, callbackFn) {
-      client.getUser(userId, function(err, res, data) {
-        console.log(res);
-        callbackFn(err, res, data);
-      });
-    };
-    apiObj.like = function(userId) {
-      client.like(userId, function(err, res, data) {
-        console.log(res);
-        if (res && res.match) {
-          apiObj.userInfo(res.match.participants[1], function(err2, res2, data2) {
-            var user = res2.results;
-            swal({
-              title: 'It\'s a match!',
-              text: 'Go send a message to ' + user.name + ' (on your phone for now)',
-              confirmButtonText: 'Nice!',
-              imageUrl: user.photos[0].processedFiles[3].url
-            });
-          });
-        }
-      });
-    };
-    apiObj.pass = function(userId) {
-      client.pass(userId, function(err, res, data) {
-        console.log(res);
-      });
-    };
-    apiObj.message = function(userId, message) {
-      client.sendMessage(userId, message, function(err, res, data) {
-        console.log(res);
-      });
-    };
-
-    return apiObj;
-  });
-
-  app.controller('SwipeController', function SwipeController($scope, $http, $timeout, $window, API) {
+  module.controller('SwipeController', function SwipeController($scope, $http, $timeout, API) {
     $scope.allPeople = [];
     $scope.peopleIndex = 0;
     $scope.showLocation = false;
@@ -371,68 +261,23 @@
 
   });
 
-  app.controller('LoginController', function LoginController($scope, $http, API) {
-    $scope.loginUrl = 'https://m.facebook.com/dialog/oauth?client_id=464891386855067&redirect_uri=https://www.facebook.com/connect/login_success.html&scope=basic_info,email,public_profile,user_about_me,user_activities,user_birthday,user_education_history,user_friends,user_interests,user_likes,user_location,user_photos,user_relationship_details&response_type=token';
-    $scope.fbAuthData = {};
-
-    $scope.startLogin = function() {
-      window.loginWindow = gui.Window.open($scope.loginUrl, {
-        title: 'Login to Facebook',
-        position: 'center',
-        width: 400,
-        height: 480,
-        focus: true
-      });
-      var interval = window.setInterval(function() {
-        if (window.loginWindow) {
-          checkForToken(window.loginWindow.window, interval);
-        }
-      }, 500);
-      window.loginWindow.on('closed', function() {
-        window.clearInterval(interval);
-        window.loginWindow = null;
-      });
-      ga_storage._trackEvent('Login', 'Login Started');
+  //helper
+  var debounce = function(func, wait, immediate) {
+    var timeout;
+    return function() {
+      var context = this, args = arguments;
+      var later = function() {
+        timeout = null;
+        if (!immediate) func.apply(context, args);
+      };
+      var callNow = immediate && !timeout;
+      clearTimeout(timeout);
+      timeout = setTimeout(later, wait);
+      if (callNow) func.apply(context, args);
     };
+  };
 
-    var tinderLogin = function() {
-      API.login($scope.fbAuthData['fb_id'], $scope.fbAuthData['access_token']);
-    };
-
-    var checkForToken = function(loginWindow, interval) {
-      if (loginWindow.closed) {
-        window.clearInterval(interval);
-      } else {
-        var url = loginWindow.document.URL;
-        var paramString = url.split("#")[1];
-        if (!!paramString) {
-          var allParam = paramString.split("&");
-          for (var i = 0; i < allParam.length; i++) {
-            var param = allParam[i].split("=");
-            $scope.fbAuthData[param[0]] = param[1];
-          }
-          loginWindow.close();
-          window.clearInterval(interval);
-          getFBUserId($scope.fbAuthData['access_token']);
-        }
-      }
-    };
-
-    var getFBUserId = function(token) {
-      var graphUrl = 'https://graph.facebook.com/me?access_token=' + token;
-      $http.get(graphUrl)
-          .success(function(data) {
-            console.log(data);
-            $scope.fbAuthData['fb_id'] = data.id;
-            tinderLogin();
-          })
-          .error(function(data) {
-            console.log(data);
-          });
-    }
-  });
-
-  app.directive('renderImagesDirective', function() {
+  module.directive('renderImagesDirective', function() {
     return function(scope, element, attrs) {
       if (scope.$last){
         scope.$emit('cardsRendered');
@@ -440,13 +285,13 @@
     };
   });
 
-  app.filter('bdayToAge', function() {
+  module.filter('bdayToAge', function() {
     return function(bday) {
       return moment.duration(moment().diff(moment(bday))).years();
     };
   });
 
-  app.filter('emoji', function() {
+  module.filter('emoji', function() {
     return function(string) {
       return twemoji.parse(string);
     };
@@ -455,7 +300,7 @@
   // based off https://github.com/doukasd/AngularJS-Components
   // a directive to auto-collapse long text
   // in elements with the "dd-text-collapse" attribute
-  app.directive('ddTextCollapse', ['$compile', function($compile) {
+  module.directive('ddTextCollapse', ['$compile', function($compile) {
 
     return {
       restrict: 'A',
@@ -542,25 +387,4 @@
     applyEl.css('opacity', confidence * (2 / 3));
     clearEl.css('opacity', 0);
   }
-
-  // helpers
-  capitalize = function (s)
-  {
-    return s[0].toUpperCase() + s.slice(1);
-  }
-
-  var debounce = function(func, wait, immediate) {
-    var timeout;
-    return function() {
-      var context = this, args = arguments;
-      var later = function() {
-        timeout = null;
-        if (!immediate) func.apply(context, args);
-      };
-      var callNow = immediate && !timeout;
-      clearTimeout(timeout);
-      timeout = setTimeout(later, wait);
-      if (callNow) func.apply(context, args);
-    };
-  };
 })();
